refactor(layout): extract Sidebar and MobileNav from MainLayout
